% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{landscapeverse_deps}
\alias{landscapeverse_deps}
\title{List all landscapeverse dependencies}
\usage{
landscapeverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
landscapeverse packages.}
}
\description{
List all landscapeverse dependencies
}
