% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{landscapeverse_packages}
\alias{landscapeverse_packages}
\title{List all packages in the landscapeverse}
\usage{
landscapeverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include landscapeverse in the list?}
}
\description{
List all packages in the landscapeverse
}
\examples{
landscapeverse_packages()
}
