% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{landscapeverse_update}
\alias{landscapeverse_update}
\title{Update landscapeverse packages}
\usage{
landscapeverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
landscapeverse packages.}
}
\description{
This will check to see if all landscapeverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
landscapeverse_update()
}
}
